holy crap theres so much dogcrap copypasta code...
