+++ packages/synapse-interface/components/StateManagedSwap/SwapInputContainer.tsx
@@ -12,142 +12,25 @@ import { usePortfolioState } from '@/slices/portfolio/hooks'
 import { updateSwapFromValue } from '@/slices/swap/reducer'
 import { useSwapState } from '@/slices/swap/hooks'
 import { ConnectStatusIndicator } from '@/components/bridgeSwap/ConnectStatusIndicator'
+import { GenericInputContainer } from '@/components/bridgeSwap/GenericInputContainer'
 
 export const SwapInputContainer = () => {
   const inputRef = useRef<HTMLInputElement>(null)
-  const { swapChainId, swapFromToken, swapToToken, swapFromValue } =
-    useSwapState()
-  const [showValue, setShowValue] = useState('')
-
-  const { balances } = usePortfolioState()
-
-  const { isConnected } = useAccount()
+  const { swapChainId, swapFromToken, swapFromValue } = useSwapState()
 
   const dispatch = useDispatch()
 
-  const tokenData = balances[swapChainId]?.find(
-    (token) => token.tokenAddress === swapFromToken?.addresses[swapChainId]
-  )
-
-  const parsedBalance = tokenData?.parsedBalance
-
-  const balance = tokenData?.balance
-
-  useEffect(() => {
-    if (
-      swapFromToken &&
-      swapFromToken.decimals[swapChainId] &&
-      stringToBigInt(swapFromValue, swapFromToken.decimals[swapChainId]) !== 0n
-    ) {
-      setShowValue(swapFromValue)
-    }
-  }, [swapFromValue, swapChainId, swapFromToken])
-
-  const handleFromValueChange = (
-    event: React.ChangeEvent<HTMLInputElement>
-  ) => {
-    const swapFromValueString: string = cleanNumberInput(event.target.value)
-    try {
-      dispatch(updateSwapFromValue(swapFromValueString))
-      setShowValue(swapFromValueString)
-    } catch (error) {
-      console.error('Invalid value for conversion to BigInteger')
-      const inputValue = event.target.value
-      const regex = /^[0-9]*[.,]?[0-9]*$/
-
-      if (regex.test(inputValue) || inputValue === '') {
-        dispatch(updateSwapFromValue(''))
-        setShowValue(inputValue)
-      }
-    }
-  }
-
-  const onMaxBalance = useCallback(() => {
-    dispatch(
-      updateSwapFromValue(
-        formatBigIntToString(balance, swapFromToken.decimals[swapChainId])
-      )
-    )
-  }, [balance, swapChainId, swapFromToken])
-
-
 
   return (
-    <div
-      data-test-id="input-container"
-      className="text-left rounded-md p-md bg-bgBase/10 ring-1 ring-white/10"
-    >
-      <div className="flex items-center justify-between mb-3">
-        <SwapChainSelector />
-        <ConnectStatusIndicator targetChainId={swapChainId} />
-      </div>
-      <div className="flex h-16 mb-2 space-x-2">
-        <div
-          className={`
-            flex flex-grow items-center justify-between
-            pl-md
-            w-full h-16
-            rounded-md
-            focus-within:bg-bgBase/10
-            border border-white/10 focus-within:border-white/40
-          `}
-        >
-          <div className="flex items-center">
-            <SwapFromTokenSelector />
-            <div className="flex flex-col justify-between ml-4">
-              <div style={{ display: 'table' }}>
-                <input
-                  ref={inputRef}
-                  pattern="^[0-9]*[.,]?[0-9]*$"
-                  disabled={false}
-                  className={`
-                    focus:outline-none
-                    focus:ring-0
-                    focus:border-none
-                    border-none
-                    bg-transparent
-                    max-w-[190px]
-                    p-0
-                    placeholder:text-white/40
-                    text-white text-opacity-80 text-xl md:text-2xl font-medium
-                  `}
-                  placeholder="0.0000"
-                  onChange={handleFromValueChange}
-                  value={showValue}
-                  name="inputRow"
-                  autoComplete="off"
-                  minLength={1}
-                  maxLength={79}
-                  style={{ display: 'table-cell', width: '100%' }}
-                />
-              </div>
-              {isConnected && (
-                <label
-                  htmlFor="inputRow"
-                  className="text-xs text-white transition-all duration-150 transform-gpu hover:text-opacity-70 hover:cursor-pointer"
-                  onClick={onMaxBalance}
-                >
-                  {parsedBalance ?? '0.0'}
-                  <span className="text-opacity-50 text-secondaryTextColor">
-                    {' '}
-                    available
-                  </span>
-                </label>
-              )}
-            </div>
-          </div>
-          <div>
-            {isConnected && (
-              <div className="m">
-                <MiniMaxButton
-                  disabled={!balance || balance === 0n}
-                  onClickBalance={onMaxBalance}
-                />
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
+    <GenericInputContainer
+      inputRef={inputRef}
+      chainId={swapChainId}
+      token={swapFromToken}
+      value={swapFromValue}
+      initialStateValue={null} // initialState.fromValue
+      dispatchUpdateFunc={(val) => dispatch(updateSwapFromValue(val))}
+      chainSelector={<SwapChainSelector />}
+      tokenSelector={<SwapFromTokenSelector />}
+    />
   )
 }
